+++ k8s/db-management.yaml
@@ -146,7 +146,7 @@ spec:
                       name: app-secrets
                       key: USER_AGENT
                 - name: RUST_LOG
-                  value: warn,sync=info
+                  value: warn,sync=info,cards_sdk=info
               volumeMounts:
                 - name: images
                   mountPath: /var/mtg_cards
