+++ packages/bot/Cargo.toml
@@ -11,7 +11,6 @@ regex = "1.11.1"
 dotenv = { version = "0.15.0", optional = true }
 serenity = "0.12.4"
 tokio = { version = "1.47.1", features = ["macros", "rt-multi-thread"] }
-sqlx = { version = "0.8.6", features = [ "runtime-tokio", "postgres", "uuid", "time" ] }
 uuid = { version = "1.18.0", features = ["v4", "serde"] }
 log = "0.4.27"
 env_logger = "0.11.8"
@@ -20,6 +19,7 @@ serde = { version = "1.0.219", features = ["derive"] }
 ron = "0.10.1"
 async-trait = "0.1.89"
 thiserror = "2.0.16"
+cards_sdk = { workspace = true }
 contracts = { workspace = true }
 fuzzy = { workspace = true }
 named_lock = { workspace = true }
@@ -27,6 +27,7 @@ normalise = { workspace = true }
 lru = "0.18.0"
 
 [dev-dependencies]
+cards_sdk = { workspace = true, features = ["test-util"] }
 mockall = "0.13.1"
 time = { version = "0.3", features = ["serde"] }
