+++ packages/cards_sdk/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "cards_sdk"
+version = "0.1.0"
+edition = "2024"
+
+[features]
+test-util = ["dep:mockall"]
+local-dev = ["dep:indicatif"]
+
+[dependencies]
+contracts = { workspace = true }
+async-trait = "0.1.89"
+sqlx = { version = "0.8.6", features = ["runtime-tokio", "postgres", "uuid", "time"] }
+uuid = "1.18.0"
+time = { version = "0.3", features = ["serde"] }
+log = "0.4.27"
+futures = "0.3.32"
+mockall = { version = "0.13.1", optional = true }
+indicatif = { version = "0.18.4", optional = true }
+
+[dev-dependencies]
+mockall = "0.13.1"
+tokio = { version = "1.47.1", features = ["macros", "rt-multi-thread"] }
+
+[lints]
+workspace = true
