+++ packages/sync/Cargo.toml
@@ -7,11 +7,12 @@ edition = "2024"
 workspace = true
 
 [features]
-local-dev = ["dotenv", "indicatif"]
+local-dev = ["dotenv", "indicatif", "cards_sdk/local-dev"]
 
 [dependencies]
 dotenv = { version = "0.15.0", optional = true }
 indicatif = { version = "0.18.4", optional = true }
+cards_sdk = { workspace = true }
 contracts = { workspace = true }
 normalise = { workspace = true }
 async-trait = "0.1.89"
@@ -21,7 +22,6 @@ serde_json = "1"
 tokio = { version = "1.47.1", features = ["macros", "rt-multi-thread"] }
 time = { version = "0.3", features = ["serde", "macros", "formatting", "parsing"] }
 uuid = { version = "1.18.0", features = ["v4"] }
-sqlx = { version = "0.8.6", features = [ "runtime-tokio", "postgres", "uuid", "time" ] }
 log = "0.4.27"
 regex = "1.12.2"
 futures = "0.3.32"
