+++ packages/sync/src/adapters/services/file_system/mod.rs
@@ -1,5 +1,5 @@
 use crate::ports::image_store::{Illustration, Image, ImageStore};
-use crate::ports::storage::CardInfo;
+use cards_sdk::CardInfo;
 use async_trait::async_trait;
 use std::env;
 use uuid::Uuid;
