+++ packages/sync/src/adapters/services/scryfall/data/set.rs
@@ -1,4 +1,4 @@
-use crate::ports::storage::Set;
+use cards_sdk::Set;
 use serde::{Deserialize, Serialize};
 use time::Date;
 use time::serde::format_description;
