+++ packages/sync/src/domain/bulk.rs
@@ -3,11 +3,11 @@ use crate::domain::utils::images::save_images;
 use crate::ports::emoji::EmojiStore;
 use crate::ports::image_store::ImageStore;
 use crate::ports::source::CardSource;
-use crate::ports::storage::Storage;
+use cards_sdk::WriteRepository;
 
 pub async fn sync(
     source: impl CardSource,
-    storage: impl Storage,
+    storage: impl WriteRepository,
     image_store: impl ImageStore,
     emoji_store: impl EmojiStore,
 ) {
