+++ packages/sync/src/ports/mod.rs
@@ -1,4 +1,3 @@
 pub mod emoji;
 pub mod image_store;
 pub mod source;
-pub mod storage;
